+++ NEWS.md
@@ -1,5 +1,7 @@
 # legendry (development version)
 
+* Fixed bug hindering `position = "inside"` placement (#42)
+
 # legendry 0.2.0
 
 This is a small feature release introducing dendrogram scales and a size guide.